Implement OFDM front-end: sync, timing, FFT, pilot extraction
Phase 2 implementation for ATSC 3.0 physical layer receiver:

- Bootstrap detector: Schmidl-Cox autocorrelation for coarse timing and
  CFO estimation at 4K bootstrap symbol
- Timing recovery: Gardner TED with 32-tap, 16-phase polyphase interpolator
- Frame sync: Superframe/subframe boundary tracking via preamble correlation
- Frequency correction: Coarse CFO from bootstrap, fine CFO from pilot
  phase slope tracking
- CP removal: Configurable cyclic prefix stripping per ATSC A/322
- FFT engine: FFTW3 backend for 8K/16K/32K transforms with wisdom caching
- Pilot extractor: Scattered pilots (PP1-PP8), continual pilots, edge pilots
  per ATSC A/322 §7.2
- Mode config: JSON-based ATSC 3.0 mode parameter loading

Includes unit tests for all components (30+ tests passing).

+++ lib/types.h
@@ -19,8 +19,8 @@ using real_t = int16_t;
 // Q1.15: 1 sign bit, 0 integer bits, 15 fractional bits
 // Range: [-1.0, 1.0) with resolution 2^-15 = 3.05e-5
 constexpr int ATSC3_Q_FRACTIONAL_BITS = 15;
-constexpr int16_t ATSC3_Q_SCALE = 1 << ATSC3_Q_FRACTIONAL_BITS;
-constexpr float ATSC3_Q_SCALE_INV = 1.0f / ATSC3_Q_SCALE;
+constexpr int32_t ATSC3_Q_SCALE = 1 << ATSC3_Q_FRACTIONAL_BITS;  // 32768
+constexpr float ATSC3_Q_SCALE_INV = 1.0f / static_cast<float>(ATSC3_Q_SCALE);
 
 // Convert float to Q1.15
 inline int16_t float_to_q15(float x) {
